requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6

# Azure SDK
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-authorization>=4.0.0
azure-mgmt-keyvault>=10.3.0
azure-mgmt-msi>=7.0.0

# Graph engine
networkx>=3.2.1

# Database
sqlalchemy>=2.0.25

# Config and rules
pyyaml>=6.0.1
pydantic>=2.5.0

# CLI
click>=8.1.7
rich>=13.7.0

# PDF export
reportlab>=4.1.0

# Authentication (Entra ID / Azure AD)
msal>=1.26.0

# HTTP client
httpx>=0.26.0
